+++ src/scirex/torch/sciml/data/__init__.py
@@ -1,0 +1,10 @@
+"""
+Data generation utilities for FNO models.
+
+This package provides tools for generating synthetic datasets for training
+Fourier Neural Operator models on various PDE problems.
+"""
+
+from .generate import DataGenerator
+
+__all__ = ['DataGenerator']
