+++ src/scirex/torch/sciml/models/fno/fno_sr.py
@@ -28,8 +28,8 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from ..config import FNOConfig
-from .fno_block import FNOBlock1d
+from ...config_sr import FNOConfig
+from .fno_block_sr import FNOBlock1d
 
 class FNO1d(nn.Module): 
     def __init__(self, config: FNOConfig):
